Add safety to lift ratchet release
The lift ratchets require a careful sequence to release before we can drive the elevator up. If the elevator is not stowed correctly we may not be able to release the ratchets. To prevent us from stalling the motors for a prolonged time we will timeout the initial move (to release tension) and will cancel the rest of the command group.